+++ packages/backend-auth/src/factory.test.ts
@@ -15,10 +15,6 @@ import {
 import { AmplifyAuth, triggerEvents } from '@aws-amplify/auth-construct-alpha';
 import { StackMetadataBackendOutputStorageStrategy } from '@aws-amplify/backend-output-storage';
 import {
-  BackendDeploymentType,
-  CDKContextKey,
-} from '@aws-amplify/platform-core';
-import {
   ConstructContainerStub,
   ImportPathVerifierStub,
   StackResolverStub,
@@ -26,12 +22,9 @@ import {
 
 const createStackAndSetContext = (): Stack => {
   const app = new App();
-  app.node.setContext('branch-name', 'testEnvName');
-  app.node.setContext('backend-id', 'testBackendId');
-  app.node.setContext(
-    CDKContextKey.DEPLOYMENT_TYPE,
-    BackendDeploymentType.BRANCH
-  );
+  app.node.setContext('amplify-backend-name', 'testEnvName');
+  app.node.setContext('amplify-backend-namespace', 'testBackendId');
+  app.node.setContext('amplify-backend-type', 'branch');
   const stack = new Stack(app);
   return stack;
 };