+++ packages/backend-storage/src/factory.test.ts
@@ -12,10 +12,6 @@ import {
   ImportPathVerifier,
 } from '@aws-amplify/plugin-types';
 import { StackMetadataBackendOutputStorageStrategy } from '@aws-amplify/backend-output-storage';
-import {
-  BackendDeploymentType,
-  CDKContextKey,
-} from '@aws-amplify/platform-core';
 import { AmplifyStorage } from '@aws-amplify/storage-construct-alpha';
 import {
   ConstructContainerStub,
@@ -25,12 +21,9 @@ import {
 
 const createStackAndSetContext = (): Stack => {
   const app = new App();
-  app.node.setContext('branch-name', 'testEnvName');
-  app.node.setContext('backend-id', 'testBackendId');
-  app.node.setContext(
-    CDKContextKey.DEPLOYMENT_TYPE,
-    BackendDeploymentType.BRANCH
-  );
+  app.node.setContext('amplify-backend-name', 'testEnvName');
+  app.node.setContext('amplify-backend-namespace', 'testBackendId');
+  app.node.setContext('amplify-backend-type', 'branch');
   const stack = new Stack(app);
   return stack;
 };