+++ packages/backend/src/engine/backend-secret/backend_secret.test.ts
@@ -2,17 +2,19 @@ import { describe, it, mock } from 'node:test';
 import assert from 'node:assert';
 import { CfnTokenBackendSecret } from './backend_secret.js';
 import { App, SecretValue, Stack } from 'aws-cdk-lib';
-import { BranchBackendIdentifier } from '@aws-amplify/platform-core';
 import { BackendSecretFetcherProviderFactory } from './backend_secret_fetcher_provider_factory.js';
 import { BackendSecretFetcherFactory } from './backend_secret_fetcher_factory.js';
-import { UniqueBackendIdentifier } from '@aws-amplify/plugin-types';
+import { BackendIdentifier } from '@aws-amplify/plugin-types';
 
-const backendId = 'testId';
-const branchName = 'testBranch';
+const namespace = 'testId';
+const name = 'testBranch';
 const testSecretName = 'testSecretName';
 const testSecretValue = 'testSecretValue';
-const uniqueBackendIdentifier: UniqueBackendIdentifier =
-  new BranchBackendIdentifier(backendId, branchName);
+const backendId: BackendIdentifier = {
+  namespace,
+  name,
+  type: 'branch',
+};
 
 void describe('BackendSecret', () => {
   const providerFactory = new BackendSecretFetcherProviderFactory();
@@ -28,7 +30,7 @@ void describe('BackendSecret', () => {
     const app = new App();
     const stack = new Stack(app);
     const secret = new CfnTokenBackendSecret(testSecretName, resourceFactory);
-    const val = secret.resolve(stack, uniqueBackendIdentifier);
+    const val = secret.resolve(stack, backendId);
     assert.deepStrictEqual(val, SecretValue.unsafePlainText(testSecretValue));
     assert.deepStrictEqual(mockGetOrCreate.mock.callCount(), 1);
   });