+++ packages/cli/src/commands/sandbox/sandbox_command.test.ts
@@ -43,7 +43,11 @@ void describe('sandbox command', () => {
 
   beforeEach(async () => {
     const sandboxFactory = new SandboxSingletonFactory(() =>
-      Promise.resolve('testBackendId')
+      Promise.resolve({
+        namespace: 'testSandboxId',
+        name: 'testSandboxName',
+        type: 'sandbox',
+      })
     );
     sandbox = await sandboxFactory.getInstance();
 
@@ -222,7 +226,11 @@ void describe('sandbox command', () => {
   void it('starts sandbox with user provided valid AWS profile', async () => {
     mockHandleProfile.mock.mockImplementationOnce(() => null);
     const sandboxFactory = new SandboxSingletonFactory(() =>
-      Promise.resolve('testBackendId')
+      Promise.resolve({
+        namespace: 'testSandboxId',
+        name: 'testSandboxName',
+        type: 'sandbox',
+      })
     );
     sandbox = await sandboxFactory.getInstance();
     sandboxStartMock = mock.method(sandbox, 'start', () => Promise.resolve());