+++ packages/deployed-backend-client/src/stack-name-resolvers/app_name_and_branch_main_stack_name_resolver.test.ts
@@ -72,6 +72,7 @@ void describe('AppNameAndBranchMainStackNameResolver', () => {
       appNameAndBranch
     );
     const result = await resolver.resolveMainStackName();
-    assert.equal(result, 'amplify-testBackendId-testBranch');
+    // eslint-disable-next-line spellcheck/spell-checker
+    assert.equal(result, 'amplify-testBackendId-testBranch-branch-51627fbc17');
   });
 });